api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.4.0
supabase==2.4.0
# Supabase 2.4.0 pins httpx<0.26
httpx==0.25.2
google-generativeai==0.7.2
aiofiles==24.1.0
PyJWT==2.9.0
python-multipart==0.0.20
PyPDF2==3.0.1
python-docx==1.1.2
